Moved model file
