+++ shared/models/socket-communication/SocketData.ts
